requirements.txt
python-dotenv
openai
Flask
Flask-Dance[google]=6.2.0
